+++ app/src/main/java/org/flyve/mdm/agent/ui/StartEnrollmentActivity.java
@@ -43,7 +43,7 @@ import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.core.supervisor.SupervisorController;
 import org.flyve.mdm.agent.core.supervisor.SupervisorModel;
 import org.flyve.mdm.agent.data.DataStorage;
-import org.flyve.mdm.agent.core.EnrollmentHelper;
+import org.flyve.mdm.agent.core.enrollment.EnrollmentHelper;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
